refactoring code

+++ php/delete_image.php
@@ -4,19 +4,22 @@
     $postdata = file_get_contents("php://input");
     $request = json_decode($postdata);
     require_once 'Connector.php';
-    $sql = "SELECT * FROM images where image_id = ".$request->imageId;
+    $sql = "SELECT * FROM images where image_id = :id";
     $stmt = $con->prepare($sql);
-    $stmt->execute();
-    $result=$stmt->fetch(PDO::FETCH_ASSOC);
-    //var_dump($sql);
-    //var_dump($result['image_url']);
-    $stmt = $con->prepare(" DELETE FROM images where image_id = :id");
     $stmt->bindParam(':id',$request->imageId);
+    $stmt->execute();
     $response = array();
-    if($stmt->execute() && unlink("../".$result['image_url'])) {
-      $response["deleted"] = true;
+    if($result = $stmt->fetch(PDO::FETCH_ASSOC)) {
+      $stmt = $con->prepare(" DELETE FROM images where image_id = :id");
+      $stmt->bindParam(':id',$result['image_id']);
+      if( unlink("../".$result['image_url']) && $stmt->execute()) {
+        $response["deleted"] = true;
+      } else {
+        $response["deleted"] = false;
+      }
     } else {
       $response["deleted"] = false;
+      //var_dump($result);
     }
     echo json_encode($response);
   }
